multiline command supports escaped newlines
