+++ lib/command/src/lib.rs
@@ -1,6 +1,9 @@
 use std::path::Path;
 
-use komodo_client::entities::{komodo_timestamp, update::Log};
+use komodo_client::{
+  entities::{komodo_timestamp, update::Log},
+  parser::parse_multiline_command,
+};
 use run_command::{async_run_command, CommandOutput};
 
 /// Parses commands out of multiline string
@@ -21,35 +24,6 @@ pub async fn run_komodo_command(
   let start_ts = komodo_timestamp();
   let output = async_run_command(&command).await;
   output_into_log(stage, command, start_ts, output)
-}
-
-/// Parses commands out of multiline string
-/// and chains them together with '&&'
-///
-/// Supports full line and end of line comments.
-///
-/// ## Example:
-/// ```sh
-/// # comments supported
-/// sh ./shell1.sh # end of line supported
-/// sh ./shell2.sh
-/// # print done
-/// echo done
-/// ```
-/// becomes
-/// ```sh
-/// sh ./shell1.sh && sh ./shell2.sh && echo done
-/// ```
-pub fn parse_multiline_command(command: impl AsRef<str>) -> String {
-  command
-    .as_ref()
-    .split('\n')
-    .map(str::trim)
-    .filter(|line| !line.is_empty() && !line.starts_with('#'))
-    .filter_map(|line| line.split(" #").next())
-    .map(str::trim)
-    .collect::<Vec<_>>()
-    .join(" && ")
 }
 
 pub fn output_into_log(
